requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --frozen --no-hashes --no-dev --output-file requirements.txt
aiosqlite==0.22.1
    # via akapen
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anthropic==0.97.0
    # via akapen
anyio==4.13.0
    # via
    #   anthropic
    #   google-genai
    #   gradio
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
audioop-lts==0.2.2 ; python_full_version >= '3.13'
    # via gradio
brotli==1.2.0
    # via gradio
certifi==2026.4.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.7
    # via requests
click==8.3.3
    # via
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==47.0.0
    # via google-auth
deprecated==1.3.1
    # via limits
distro==1.9.0
    # via
    #   anthropic
    #   google-genai
    #   openai
docstring-parser==0.18.0
    # via anthropic
fastapi==0.136.1
    # via
    #   akapen
    #   gradio
filelock==3.29.0
    # via huggingface-hub
fsspec==2026.4.0
    # via
    #   gradio-client
    #   huggingface-hub
google-auth==2.50.0
    # via google-genai
google-genai==1.74.0
    # via akapen
gradio==6.14.0
    # via akapen
gradio-client==2.5.0
    # via
    #   gradio
    #   hf-gradio
groovy==0.1.2
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-gradio==0.4.1
    # via gradio
hf-xet==1.4.3 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   akapen
    #   anthropic
    #   google-genai
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   safehttpx
huggingface-hub==1.13.0
    # via
    #   gradio
    #   gradio-client
idna==3.13
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via gradio
jiter==0.14.0
    # via
    #   anthropic
    #   openai
limits==5.8.0
    # via slowapi
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.4.4
    # via
    #   gradio
    #   pandas
openai==1.109.1
    # via akapen
orjson==3.11.8
    # via gradio
packaging==26.2
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   limits
pandas==3.0.2
    # via gradio
pillow==12.2.0
    # via
    #   akapen
    #   gradio
prometheus-client==0.25.0
    # via akapen
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.13.3
    # via
    #   akapen
    #   anthropic
    #   fastapi
    #   google-genai
    #   gradio
    #   openai
pydantic-core==2.46.3
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.20.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via
    #   akapen
    #   uvicorn
python-multipart==0.0.27
    # via
    #   akapen
    #   gradio
pytz==2026.1.post1
    # via gradio
pyyaml==6.0.3
    # via
    #   gradio
    #   huggingface-hub
    #   uvicorn
requests==2.33.1
    # via
    #   google-auth
    #   google-genai
rich==15.0.0
    # via typer
safehttpx==0.1.7
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
slowapi==0.1.9
    # via akapen
sniffio==1.3.1
    # via
    #   anthropic
    #   google-genai
    #   openai
starlette==1.0.0
    # via
    #   fastapi
    #   gradio
tenacity==9.1.4
    # via google-genai
tomlkit==0.14.0
    # via gradio
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   openai
typer==0.25.1
    # via
    #   gradio
    #   hf-gradio
    #   huggingface-hub
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   fastapi
    #   google-genai
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   limits
    #   openai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
tzdata==2026.2 ; sys_platform == 'emscripten' or sys_platform == 'win32'
    # via pandas
urllib3==2.6.3
    # via requests
uvicorn==0.46.0
    # via
    #   akapen
    #   gradio
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via
    #   google-genai
    #   uvicorn
wrapt==2.1.2
    # via deprecated
